+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Bump shinylive assets dependency to 0.1.1.
 
+* Use `{httpuv}` to serve static folder instead of plumber. (#40)
+* Use `{httr2}` to download assets from GitHub releases. (@dgkf #30, #39)
+
 # shinylive 0.1.0
 
 * Initial CRAN submission.